+++ modules/core/src/component/dynamic-form-control.component.ts
@@ -4,6 +4,7 @@ import {DynamicFormControlModel} from "../model/dynamic-form-control.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX} from "../model/checkbox/dynamic-checkbox.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP} from "../model/checkbox/dynamic-checkbox-group.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP} from "../model/radio/dynamic-radio-group.model";
+import {isDefined} from "../utils";
 
 export abstract class DynamicFormControlComponent implements OnInit {
 
@@ -16,8 +17,7 @@ export abstract class DynamicFormControlComponent implements OnInit {
 
     incompatibilities: Array<string> = [];
 
-    constructor() {
-    }
+    constructor() {}
 
     ngOnInit() {
 
@@ -25,7 +25,16 @@ export abstract class DynamicFormControlComponent implements OnInit {
             throw new Error(`Control ${this.model.id} of type ${this.model.type} is not supported by ${this.type} UI package.`);
         }
 
+        if (!isDefined(this.model)) {
+            throw new Error(`no model input defined for DynamicFormControlComponent`);
+        }
+
+        if (!isDefined(this.controlGroup)) {
+            throw new Error(`no controlGroup input defined for DynamicFormControlComponent`);
+        }
+
         this.control = <FormControl> this.controlGroup.get(this.model.id);
+
         //@exclude
         this.control.valueChanges.subscribe((value: any) => {
             console.log(this.model.id + " field changed to: ", value, typeof value, this.control.valid);