+++ modules/core/src/service/dynamic-form.service.ts
@@ -8,7 +8,12 @@ import {
     DYNAMIC_FORM_CONTROL_TYPE_ARRAY
 } from "../model/form-array/dynamic-form-array.model";
 import {DYNAMIC_FORM_CONTROL_TYPE_GROUP, DynamicFormGroupModel} from "../model/form-group/dynamic-form-group.model";
-import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP} from "../model/checkbox/dynamic-checkbox-group.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP, DynamicCheckboxGroupModel} from "../model/checkbox/dynamic-checkbox-group.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX, DynamicCheckboxModel,} from "../model/checkbox/dynamic-checkbox.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_INPUT, DynamicInputModel} from "../model/input/dynamic-input.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP, DynamicRadioGroupModel} from "../model/radio/dynamic-radio-group.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_SELECT, DynamicSelectModel} from "../model/select/dynamic-select.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_TEXTAREA, DynamicTextAreaModel} from "../model/textarea/dynamic-textarea.model";
 
 @Injectable()
 export class DynamicFormService {
@@ -65,7 +70,6 @@ export class DynamicFormService {
         return this.formBuilder.group(formGroup, groupExtra);
     }
 
-
     createFormArrayGroup(dynamicFormArrayModel: DynamicFormArrayModel): FormGroup {
 
         return this.createFormGroup(dynamicFormArrayModel.addGroup().group);
@@ -96,5 +100,56 @@ export class DynamicFormService {
 
     findById(id: string, group: Array<DynamicFormControlModel>): DynamicFormControlModel {
         return group.find(controlModel => controlModel.id === id);
+    }
+
+    fromJSON(json: Array<Object>): Array<DynamicFormControlModel> {
+
+        let formModel: Array<DynamicFormControlModel> = [];
+
+        json.forEach(object => {
+
+            switch (object["type"]) {
+
+                case DYNAMIC_FORM_CONTROL_TYPE_ARRAY:
+                    object["createGroup"] = () => this.fromJSON(object["originGroup"]);
+                    formModel.push(new DynamicFormArrayModel(object, object["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX:
+                    formModel.push(new DynamicCheckboxModel(object, object["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP:
+                    object["group"] = this.fromJSON(object["group"]);
+                    formModel.push(new DynamicCheckboxGroupModel(object, object["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_GROUP:
+                    object["group"] = this.fromJSON(object["group"]);
+                    formModel.push(new DynamicFormGroupModel(object, object["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_INPUT:
+                    formModel.push(new DynamicInputModel(object, object["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP:
+                    formModel.push(new DynamicRadioGroupModel(object, object["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_SELECT:
+                    formModel.push(new DynamicSelectModel(object, object["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_TEXTAREA:
+                    formModel.push(new DynamicTextAreaModel(object, object["cls"]));
+                    break;
+
+                default:
+                    throw new Error(`no known form control type defined on JSON object`);
+            }
+        });
+
+        return formModel;
     }
 }