+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# beta.17
+
+### **Breaking Changes** 
+
+* none
+
+### **Bugfixes** 
+
+* none
+
+### **Features**
+ 
+* Migration to Angular 2 RC.7 (closes [#100](https://github.com/udos86/ng2-dynamic-forms/issues/100)) 
+* `fromJSON()` function added to `DynamicFormService` (closes [#99](https://github.com/udos86/ng2-dynamic-forms/issues/99))
+
+
 # beta.16
 
 ### **Breaking Changes**